requirements.txt
pydub
tqdm
azure-cognitiveservices-speech
python-dotenv